Improved code

+++ Python/Keep a specific ammount of log files in a folder deleting old ones.py
@@ -2,24 +2,39 @@ import logging
 from pathlib import Path
 import os
 
-def removeOldLogFiles():
-	"""Remove old logs based on limit in config"""
-	logging.debug(f'Deleting old log files.')
-	logs_to_keep = 30
-		
-	logging.debug(f'Getting list of log files to delete.')
-	old_logs = list(sorted(Path('logs').iterdir(), key = os.path.getctime, reverse = False))[1:] #Source: https://stackoverflow.com/a/539024/16383324
-	old_logs.reverse()
-	old_logs = old_logs[logs_to_keep:]
-	logging.debug(f'Found {len(old_logs)} log(s) to remove.')
+def remove_old_log_files(logs_to_keep: int):
+	"""
+	Removes old log files in the 'logs' directory, keeping a specified number of logs.
 	
-	if len(old_logs) > 1:
-		old_logs_deleted = 0
-		for old_log in old_logs:
-			try:
-				os.remove(old_log)
-				logging.debug(f'Deleted "{old_log}".')
-				old_logs_deleted += 1
-			except Exception as e:
-				logging.warn(f'Cannot delete old log file "old_log". {repr(e)}')
-		logging.debug(f'Deleted {old_logs_deleted} old log(s).')
+	Args:
+		logs_to_keep (int): The number of most recent log files to keep.
+	"""
+
+	logging.info('Deleting old log files.')
+	try:
+		logs_dir = Path('logs')
+		if not logs_dir.exists():
+			logging.exception(f'Logs directory "{logs_dir}" not found.')
+			raise
+		
+		logging.debug('Getting list of log files to delete.')
+		old_logs = list(sorted(logs_dir.iterdir(), key=os.path.getctime, reverse=False))[1:]
+		old_logs.reverse()
+		old_logs = old_logs[logs_to_keep:]
+		logging.info(f'Found {len(old_logs)} log(s) to remove.')
+		
+		if len(old_logs) > 1:
+			old_logs_deleted = 0
+			for old_log in old_logs:
+				try:
+					os.remove(old_log)
+					logging.info(f'Deleted "{old_log}".')
+					old_logs_deleted += 1
+				except FileNotFoundError:
+					logging.warning(f'Cannot delete old log file "{old_log}". File not found.')
+				except Exception as e:
+					logging.warning(f'Cannot delete old log file "{old_log}". {repr(e)}')
+			logging.info(f'Deleted {old_logs_deleted} old log(s).')
+	
+	except Exception as e:
+		logging.warning(f'An error occurred while removing old log files: {repr(e)}')
